Move access control tests out of smoke tests
These tests are for the engine and are not connector smoke tests.
