+++ presto-base-jdbc/src/test/java/com/facebook/presto/plugin/jdbc/TestJdbcIntegrationSmokeTest.java
@@ -34,10 +34,4 @@ public class TestJdbcIntegrationSmokeTest
         // tablesample system doesn't work reliably with jdbc connector
         // because it generates a single split
     }
-
-    @Override
-    public void testViewAccessControl()
-    {
-        // jdbc connector does not support views
-    }
 }
