+++ presto-postgresql/src/test/java/com/facebook/presto/plugin/postgresql/TestPostgreSqlIntegrationSmokeTest.java
@@ -55,12 +55,6 @@ public class TestPostgreSqlIntegrationSmokeTest
         // because it generates a single split
     }
 
-    @Override
-    public void testViewAccessControl()
-    {
-        // jdbc connector does not support views
-    }
-
     @AfterClass(alwaysRun = true)
     public final void destroy()
             throws IOException
